requirements.txt
python-dotenv~=1.0.1
SQLAlchemy~=2.0.36
uvicorn~=0.32.0
fastapi~=0.115.2
pydantic~=2.9.2
openai~=1.51.2
pytest~=8.3.3
httpx~=0.27.2
asyncpg
